+++ railties/test/generators/app_generator_test.rb
@@ -55,6 +55,7 @@ DEFAULT_APP_FILES = %w(
   config/environments/test.rb
   config/initializers/assets.rb
   config/initializers/content_security_policy.rb
+  config/initializers/debugger.rb
   config/initializers/enable_yjit.rb
   config/initializers/filter_parameter_logging.rb
   config/initializers/inflections.rb